Tests/LibWeb/Text/expected/display_list/inline-selection-paint-invalidation.txt

New FillRect commands after selection: 4
